Fixes for dbCalendar get rid of duplicate id_trio entries. Improvements for stringRecord and stringTable. More type safety for data (not only static_casts are used more often, but checks for limits and some argument types for functions were altered). The ID fields in DBRecord were converted all to unsigned int (as they were already being used, but several functions were not taking advantage of it). A dumb crashRestore bug was fixed. The report generator now works for all types of report and it fixes demanded several fixes elsewhere, including the aforementioned dbCalendar fixes.

+++ VivenciaManager3/textdb.cpp
@@ -391,7 +391,7 @@ bool configFile::setFieldValue ( const uint field_index, const QString& value )
 		{
 			section_info->values[field_index] = value;
 			const int line_len ( section_info->fields.at ( field_index ).count () + value.count () );
-			if ( ( signed ) m_buffersize < line_len )
+			if ( static_cast<int>(m_buffersize) < line_len )
 				m_buffersize = line_len + 5;
 			m_needsaving = true;
 			return true;
